ser/value_tests: Move macros tests to macros_tests.rs
