+++ scylla-cql/src/serialize/value_tests.rs
@@ -1,4 +1,3 @@
-use crate::SerializeValue;
 use crate::frame::response::result::{CollectionType, ColumnType, NativeType, UserDefinedType};
 use crate::serialize::value::{
     BuiltinSerializationError, BuiltinSerializationErrorKind, BuiltinTypeCheckError,
@@ -733,894 +732,6 @@ fn test_cql_value_udt_errors() {
             expected: &[ColumnType::Native(NativeType::Int)],
         }
     );
-}
-
-// Do not remove. It's not used in tests but we keep it here to check that
-// we properly ignore warnings about unused variables, unnecessary `mut`s
-// etc. that usually pop up when generating code for empty structs.
-#[expect(unused)]
-#[derive(SerializeValue)]
-#[scylla(crate = crate)]
-struct TestUdtWithNoFields {}
-
-#[derive(SerializeValue, Debug, PartialEq, Eq, Default)]
-#[scylla(crate = crate)]
-struct TestUdtWithFieldSorting {
-    a: String,
-    b: i32,
-    c: Vec<i64>,
-}
-
-#[test]
-fn test_udt_serialization_with_field_sorting_correct_order() {
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let reference = do_serialize(
-        CqlValue::UserDefinedType {
-            keyspace: "ks".to_string(),
-            name: "typ".to_string(),
-            fields: vec![
-                (
-                    "a".to_string(),
-                    Some(CqlValue::Text(String::from("Ala ma kota"))),
-                ),
-                ("b".to_string(), Some(CqlValue::Int(42))),
-                (
-                    "c".to_string(),
-                    Some(CqlValue::List(vec![
-                        CqlValue::BigInt(1),
-                        CqlValue::BigInt(2),
-                        CqlValue::BigInt(3),
-                    ])),
-                ),
-            ],
-        },
-        &typ,
-    );
-    let udt = do_serialize(
-        TestUdtWithFieldSorting {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-            c: vec![1, 2, 3],
-        },
-        &typ,
-    );
-
-    assert_eq!(reference, udt);
-}
-
-#[test]
-fn test_udt_serialization_with_field_sorting_incorrect_order() {
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                // Two first columns are swapped
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let reference = do_serialize(
-        CqlValue::UserDefinedType {
-            keyspace: "ks".to_string(),
-            name: "typ".to_string(),
-            fields: vec![
-                // FIXME: UDTs in CqlValue should also honor the order
-                // For now, it's swapped here as well
-                ("b".to_string(), Some(CqlValue::Int(42))),
-                (
-                    "a".to_string(),
-                    Some(CqlValue::Text(String::from("Ala ma kota"))),
-                ),
-                (
-                    "c".to_string(),
-                    Some(CqlValue::List(vec![
-                        CqlValue::BigInt(1),
-                        CqlValue::BigInt(2),
-                        CqlValue::BigInt(3),
-                    ])),
-                ),
-            ],
-        },
-        &typ,
-    );
-    let udt = do_serialize(
-        TestUdtWithFieldSorting {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-            c: vec![1, 2, 3],
-        },
-        &typ,
-    );
-
-    assert_eq!(reference, udt);
-}
-
-#[test]
-fn test_udt_serialization_with_missing_rust_fields_at_end() {
-    let udt = TestUdtWithFieldSorting::default();
-
-    let typ_normal = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let typ_unexpected_field = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-                // Unexpected fields
-                ("d".into(), ColumnType::Native(NativeType::Counter)),
-                ("e".into(), ColumnType::Native(NativeType::Counter)),
-            ],
-        }),
-    };
-
-    let result_normal = do_serialize(&udt, &typ_normal);
-    let result_additional_field = do_serialize(&udt, &typ_unexpected_field);
-
-    assert_eq!(result_normal, result_additional_field);
-}
-
-#[derive(SerializeValue, Debug, PartialEq, Default)]
-#[scylla(crate = crate)]
-struct TestUdtWithFieldSorting2 {
-    a: String,
-    b: i32,
-    d: Option<Counter>,
-    c: Vec<i64>,
-}
-
-#[derive(SerializeValue, Debug, PartialEq, Default)]
-#[scylla(crate = crate)]
-struct TestUdtWithFieldSorting3 {
-    a: String,
-    b: i32,
-    d: Option<Counter>,
-    e: Option<f32>,
-    c: Vec<i64>,
-}
-
-#[test]
-fn test_udt_serialization_with_missing_rust_field_in_middle() {
-    let udt = TestUdtWithFieldSorting::default();
-    let udt2 = TestUdtWithFieldSorting2::default();
-    let udt3 = TestUdtWithFieldSorting3::default();
-
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                // Unexpected fields
-                ("d".into(), ColumnType::Native(NativeType::Counter)),
-                ("e".into(), ColumnType::Native(NativeType::Float)),
-                // Remaining normal field
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let result_1 = do_serialize(udt, &typ);
-    let result_2 = do_serialize(udt2, &typ);
-    let result_3 = do_serialize(udt3, &typ);
-
-    assert_eq!(result_1, result_2);
-    assert_eq!(result_2, result_3);
-}
-
-#[test]
-fn test_udt_serialization_failing_type_check() {
-    let typ_not_udt = ColumnType::Native(NativeType::Ascii);
-    let udt = TestUdtWithFieldSorting::default();
-    let mut data = Vec::new();
-
-    let err = udt
-        .serialize(&typ_not_udt, CellWriter::new(&mut data))
-        .unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::NotUdt)
-    );
-
-    let typ_without_c = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                // Last field is missing
-            ],
-        }),
-    };
-
-    let err = udt
-        .serialize(&typ_without_c, CellWriter::new(&mut data))
-        .unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::ValueMissingForUdtField { .. })
-    );
-
-    let typ_wrong_type = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                ("c".into(), ColumnType::Native(NativeType::TinyInt)), // Wrong column type
-            ],
-        }),
-    };
-
-    let err = udt
-        .serialize(&typ_wrong_type, CellWriter::new(&mut data))
-        .unwrap_err();
-    let err = err.downcast_ref::<BuiltinSerializationError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinSerializationErrorKind::UdtError(
-            UdtSerializationErrorKind::FieldSerializationFailed { .. }
-        )
-    );
-}
-
-#[derive(SerializeValue)]
-#[scylla(crate = crate)]
-struct TestUdtWithGenerics<'a, T: SerializeValue> {
-    a: &'a str,
-    b: T,
-}
-
-#[test]
-fn test_udt_serialization_with_generics() {
-    // A minimal smoke test just to test that it works.
-    fn check_with_type<T: SerializeValue>(typ: ColumnType, t: T, cql_t: CqlValue) {
-        let typ = ColumnType::UserDefinedType {
-            frozen: false,
-            definition: Arc::new(UserDefinedType {
-                name: "typ".into(),
-                keyspace: "ks".into(),
-                field_types: vec![
-                    ("a".into(), ColumnType::Native(NativeType::Text)),
-                    ("b".into(), typ),
-                ],
-            }),
-        };
-        let reference = do_serialize(
-            CqlValue::UserDefinedType {
-                keyspace: "ks".to_string(),
-                name: "typ".to_string(),
-                fields: vec![
-                    (
-                        "a".to_string(),
-                        Some(CqlValue::Text(String::from("Ala ma kota"))),
-                    ),
-                    ("b".to_string(), Some(cql_t)),
-                ],
-            },
-            &typ,
-        );
-        let udt = do_serialize(
-            TestUdtWithGenerics {
-                a: "Ala ma kota",
-                b: t,
-            },
-            &typ,
-        );
-        assert_eq!(reference, udt);
-    }
-
-    check_with_type(
-        ColumnType::Native(NativeType::Int),
-        123_i32,
-        CqlValue::Int(123_i32),
-    );
-    check_with_type(
-        ColumnType::Native(NativeType::Double),
-        123_f64,
-        CqlValue::Double(123_f64),
-    );
-}
-
-#[derive(SerializeValue, Debug, PartialEq, Eq, Default)]
-#[scylla(crate = crate, flavor = "enforce_order")]
-struct TestUdtWithEnforcedOrder {
-    a: String,
-    b: i32,
-    c: Vec<i64>,
-}
-
-#[test]
-fn test_udt_serialization_with_enforced_order_correct_order() {
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let reference = do_serialize(
-        CqlValue::UserDefinedType {
-            keyspace: "ks".to_string(),
-            name: "typ".to_string(),
-            fields: vec![
-                (
-                    "a".to_string(),
-                    Some(CqlValue::Text(String::from("Ala ma kota"))),
-                ),
-                ("b".to_string(), Some(CqlValue::Int(42))),
-                (
-                    "c".to_string(),
-                    Some(CqlValue::List(vec![
-                        CqlValue::BigInt(1),
-                        CqlValue::BigInt(2),
-                        CqlValue::BigInt(3),
-                    ])),
-                ),
-            ],
-        },
-        &typ,
-    );
-    let udt = do_serialize(
-        TestUdtWithEnforcedOrder {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-            c: vec![1, 2, 3],
-        },
-        &typ,
-    );
-
-    assert_eq!(reference, udt);
-}
-
-#[test]
-fn test_udt_serialization_with_enforced_order_additional_field() {
-    let udt = TestUdtWithEnforcedOrder::default();
-
-    let typ_normal = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let typ_unexpected_field = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-                // Unexpected field
-                ("d".into(), ColumnType::Native(NativeType::Counter)),
-            ],
-        }),
-    };
-
-    let result_normal = do_serialize(&udt, &typ_normal);
-    let result_additional_field = do_serialize(&udt, &typ_unexpected_field);
-
-    assert_eq!(result_normal, result_additional_field);
-}
-
-#[test]
-fn test_udt_serialization_with_enforced_order_failing_type_check() {
-    let typ_not_udt = ColumnType::Native(NativeType::Ascii);
-    let udt = TestUdtWithEnforcedOrder::default();
-
-    let mut data = Vec::new();
-
-    let err = <_ as SerializeValue>::serialize(&udt, &typ_not_udt, CellWriter::new(&mut data))
-        .unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::NotUdt)
-    );
-
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                // Two first columns are swapped
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let err = <_ as SerializeValue>::serialize(&udt, &typ, CellWriter::new(&mut data)).unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::FieldNameMismatch { .. })
-    );
-
-    let typ_without_c = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                // Last field is missing
-            ],
-        }),
-    };
-
-    let err = <_ as SerializeValue>::serialize(&udt, &typ_without_c, CellWriter::new(&mut data))
-        .unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::ValueMissingForUdtField { .. })
-    );
-
-    let typ_unexpected_field = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                ("c".into(), ColumnType::Native(NativeType::TinyInt)), // Wrong column type
-            ],
-        }),
-    };
-
-    let err =
-        <_ as SerializeValue>::serialize(&udt, &typ_unexpected_field, CellWriter::new(&mut data))
-            .unwrap_err();
-    let err = err.downcast_ref::<BuiltinSerializationError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinSerializationErrorKind::UdtError(
-            UdtSerializationErrorKind::FieldSerializationFailed { .. }
-        )
-    );
-}
-
-#[derive(SerializeValue, Debug)]
-#[scylla(crate = crate)]
-struct TestUdtWithFieldRename {
-    a: String,
-    #[scylla(rename = "x")]
-    b: i32,
-}
-
-#[derive(SerializeValue, Debug)]
-#[scylla(crate = crate, flavor = "enforce_order")]
-struct TestUdtWithFieldRenameAndEnforceOrder {
-    a: String,
-    #[scylla(rename = "x")]
-    b: i32,
-}
-
-#[test]
-fn test_udt_serialization_with_field_rename() {
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("x".into(), ColumnType::Native(NativeType::Int)),
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-            ],
-        }),
-    };
-
-    let mut reference = Vec::new();
-    // Total length of the struct is 23
-    reference.extend_from_slice(&23i32.to_be_bytes());
-    // Field 'x'
-    reference.extend_from_slice(&4i32.to_be_bytes());
-    reference.extend_from_slice(&42i32.to_be_bytes());
-    // Field 'a'
-    reference.extend_from_slice(&("Ala ma kota".len() as i32).to_be_bytes());
-    reference.extend_from_slice("Ala ma kota".as_bytes());
-
-    let udt = do_serialize(
-        TestUdtWithFieldRename {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-        },
-        &typ,
-    );
-
-    assert_eq!(reference, udt);
-}
-
-#[test]
-fn test_udt_serialization_with_field_rename_and_enforce_order() {
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("x".into(), ColumnType::Native(NativeType::Int)),
-            ],
-        }),
-    };
-
-    let mut reference = Vec::new();
-    // Total length of the struct is 23
-    reference.extend_from_slice(&23i32.to_be_bytes());
-    // Field 'a'
-    reference.extend_from_slice(&("Ala ma kota".len() as i32).to_be_bytes());
-    reference.extend_from_slice("Ala ma kota".as_bytes());
-    // Field 'x'
-    reference.extend_from_slice(&4i32.to_be_bytes());
-    reference.extend_from_slice(&42i32.to_be_bytes());
-
-    let udt = do_serialize(
-        TestUdtWithFieldRenameAndEnforceOrder {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-        },
-        &typ,
-    );
-
-    assert_eq!(reference, udt);
-}
-
-#[expect(unused)]
-#[derive(SerializeValue, Debug)]
-#[scylla(crate = crate, flavor = "enforce_order", skip_name_checks)]
-struct TestUdtWithSkippedNameChecks {
-    a: String,
-    b: i32,
-}
-
-#[test]
-fn test_udt_serialization_with_skipped_name_checks() {
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("x".into(), ColumnType::Native(NativeType::Int)),
-            ],
-        }),
-    };
-
-    let mut reference = Vec::new();
-    // Total length of the struct is 23
-    reference.extend_from_slice(&23i32.to_be_bytes());
-    // Field 'a'
-    reference.extend_from_slice(&("Ala ma kota".len() as i32).to_be_bytes());
-    reference.extend_from_slice("Ala ma kota".as_bytes());
-    // Field 'x'
-    reference.extend_from_slice(&4i32.to_be_bytes());
-    reference.extend_from_slice(&42i32.to_be_bytes());
-
-    let udt = do_serialize(
-        TestUdtWithFieldRenameAndEnforceOrder {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-        },
-        &typ,
-    );
-
-    assert_eq!(reference, udt);
-}
-
-#[derive(SerializeValue, Debug, PartialEq, Eq, Default)]
-#[scylla(crate = crate, forbid_excess_udt_fields)]
-struct TestStrictUdtWithFieldSorting {
-    a: String,
-    b: i32,
-    c: Vec<i64>,
-}
-
-#[test]
-fn test_strict_udt_with_field_sorting_rejects_additional_field() {
-    let udt = TestStrictUdtWithFieldSorting::default();
-    let mut data = Vec::new();
-
-    let typ_unexpected_field = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-                // Unexpected field
-                ("d".into(), ColumnType::Native(NativeType::Counter)),
-            ],
-        }),
-    };
-
-    let err = udt
-        .serialize(&typ_unexpected_field, CellWriter::new(&mut data))
-        .unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::NoSuchFieldInUdt { .. })
-    );
-
-    let typ_unexpected_field_middle = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                // Unexpected field
-                ("b_c".into(), ColumnType::Native(NativeType::Counter)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let err = udt
-        .serialize(&typ_unexpected_field_middle, CellWriter::new(&mut data))
-        .unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::NoSuchFieldInUdt { .. })
-    );
-}
-
-#[derive(SerializeValue, Debug, PartialEq, Eq, Default)]
-#[scylla(crate = crate, flavor = "enforce_order", forbid_excess_udt_fields)]
-struct TestStrictUdtWithEnforcedOrder {
-    a: String,
-    b: i32,
-    c: Vec<i64>,
-}
-
-#[test]
-fn test_strict_udt_with_enforced_order_rejects_additional_field() {
-    let udt = TestStrictUdtWithEnforcedOrder::default();
-    let mut data = Vec::new();
-
-    let typ_unexpected_field = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-                // Unexpected field
-                ("d".into(), ColumnType::Native(NativeType::Counter)),
-            ],
-        }),
-    };
-
-    let err =
-        <_ as SerializeValue>::serialize(&udt, &typ_unexpected_field, CellWriter::new(&mut data))
-            .unwrap_err();
-    let err = err.downcast_ref::<BuiltinTypeCheckError>().unwrap();
-    assert_matches!(
-        err.kind,
-        BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::NoSuchFieldInUdt { .. })
-    );
-}
-
-#[derive(SerializeValue, Debug)]
-#[scylla(crate = crate, flavor = "enforce_order", skip_name_checks)]
-struct TestUdtWithSkippedFields {
-    a: String,
-    b: i32,
-    #[scylla(skip)]
-    #[expect(dead_code)]
-    skipped: Vec<String>,
-    c: Vec<i64>,
-}
-
-#[test]
-fn test_row_serialization_with_skipped_field() {
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![
-                ("a".into(), ColumnType::Native(NativeType::Text)),
-                ("b".into(), ColumnType::Native(NativeType::Int)),
-                (
-                    "c".into(),
-                    ColumnType::Collection {
-                        frozen: false,
-                        typ: CollectionType::List(Box::new(ColumnType::Native(NativeType::BigInt))),
-                    },
-                ),
-            ],
-        }),
-    };
-
-    let reference = do_serialize(
-        TestUdtWithFieldSorting {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-            c: vec![1, 2, 3],
-        },
-        &typ,
-    );
-    let row = do_serialize(
-        TestUdtWithSkippedFields {
-            a: "Ala ma kota".to_owned(),
-            b: 42,
-            skipped: vec!["abcd".to_owned(), "efgh".to_owned()],
-            c: vec![1, 2, 3],
-        },
-        &typ,
-    );
-
-    assert_eq!(reference, row);
-}
-
-#[test]
-fn test_udt_with_non_rust_ident() {
-    #[derive(SerializeValue, Debug)]
-    #[scylla(crate = crate)]
-    struct UdtWithNonRustIdent {
-        #[scylla(rename = "a$a")]
-        a: i32,
-    }
-
-    let typ = ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "typ".into(),
-            keyspace: "ks".into(),
-            field_types: vec![("a$a".into(), ColumnType::Native(NativeType::Int))],
-        }),
-    };
-    let value = UdtWithNonRustIdent { a: 42 };
-
-    let mut reference = Vec::new();
-    // Total length of the struct
-    reference.extend_from_slice(&8i32.to_be_bytes());
-    // Field 'a'
-    reference.extend_from_slice(&(std::mem::size_of_val(&value.a) as i32).to_be_bytes());
-    reference.extend_from_slice(&value.a.to_be_bytes());
-
-    let udt = do_serialize(value, &typ);
-
-    assert_eq!(reference, udt);
 }
 
 // Tests migrated from old frame/value_tests.rs file
